+++ jest-bundle.config.js
@@ -7,6 +7,6 @@ module.exports = {
         '^.+\\.jsx?$': 'babel-jest',
     },
     transformIgnorePatterns: [
-        'node_modules/(?!(minterjs-util|minterjs-tx|buffer-es6)/)',
+        'node_modules/(?!(buffer-es6)/)',
     ],
 };